# openapi.yaml
openapi: 3.0.3
info:
  title: Ordryn / GoTodo API
  description: |
    JSON API for Ordryn (GoTodo). Self-host with `GOTODO_MODE=full` (UI + API) or
    `GOTODO_MODE=api` (API only). Redis is required for `/api/v1` authentication and rate limiting.

    Auth: session cookie (SPA login/register) or `Authorization: Bearer <api_key>`.
  version: 1.0.0
  license:
    name: See LICENSE in repository
servers:
  - url: /
    description: Same-origin (BASE_PATH may prefix these paths)
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register and start a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created; Set-Cookie session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/APIError"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login and start a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated; Set-Cookie session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/auth/username-available:
    get:
      tags: [Auth]
      summary: Check whether a username is available
      security: []
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  available:
                    type: boolean
                  valid:
                    type: boolean
                  message:
                    type: string
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, confirm_email]
              properties:
                email:
                  type: string
                  format: email
                confirm_email:
                  type: string
                  format: email
      responses:
        "200":
          description: Request accepted (always returns ok when valid)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/reset-password:
    get:
      tags: [Auth]
      summary: Validate a password reset token
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  email:
                    type: string
                    format: email
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [Auth]
      summary: Set a new password using a reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, token, new_password, confirm_password]
              properties:
                id:
                  type: string
                token:
                  type: string
                new_password:
                  type: string
                  format: password
                confirm_password:
                  type: string
                  format: password
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/auth/device/code:
    post:
      tags: [Device Auth]
      summary: Start device authorization (Android SSO)
      description: |
        Optional `redirect_uri` binds a native deep link (allowlisted `ordryn://auth-complete`)
        to this authorization. After the user approves or denies in the browser, the SPA
        navigates to that URI so the Android app can resume.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                client_name:
                  type: string
                redirect_uri:
                  type: string
                  description: Optional native deep link; must be ordryn://auth-complete
                  example: ordryn://auth-complete
      responses:
        "200":
          description: Device codes issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceCodeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/auth/device/token:
    post:
      tags: [Device Auth]
      summary: Poll device authorization for API key
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_code, grant_type]
              properties:
                device_code:
                  type: string
                grant_type:
                  type: string
                  description: Must be urn:ietf:params:oauth:grant-type:device_code
      responses:
        "200":
          description: Approved; returns api_key once
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key:
                    type: string
                  key_prefix:
                    type: string
                  key_name:
                    type: string
        "400":
          description: Pending, denied, expired, or invalid (OAuth-style error body)
  /api/v1/auth/device/status:
    get:
      tags: [Device Auth]
      summary: Look up a pending device authorization by user_code
      security: []
      parameters:
        - name: user_code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device auth status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceStatus"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/auth/device/approve:
    post:
      tags: [Device Auth]
      summary: Approve a pending device authorization
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_code]
              properties:
                user_code:
                  type: string
      responses:
        "200":
          description: Approved; includes redirect_uri when the device requested one
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceDecisionResponse"
  /api/v1/auth/device/deny:
    post:
      tags: [Device Auth]
      summary: Deny a pending device authorization
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_code]
              properties:
                user_code:
                  type: string
      responses:
        "200":
          description: Denied; includes redirect_uri when the device requested one
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceDecisionResponse"
  /api/v1/me:
    get:
      tags: [Profile]
      summary: Current user profile
      description: |
        Unauthenticated session probes return HTTP 200 with a JSON `null` body
        (not 401), so SPA bootstrap works behind proxies that rewrite 401 to HTML.
        Invalid Bearer tokens still return 401.
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Profile, or null when no session
          content:
            application/json:
              schema:
                nullable: true
                allOf:
                  - $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [Profile]
      summary: Update profile fields
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatch"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/me/username:
    post:
      tags: [Profile]
      summary: One-time username claim for migrated accounts
      description: |
        Allowed only when `username_change_available` is true. Sets the username
        and clears the free-change flag. New registrations do not get this flag.
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_name]
              properties:
                user_name:
                  type: string
                  minLength: 3
                  maxLength: 32
                  pattern: '^[A-Za-z0-9_]+$'
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
  /api/v1/me/password:
    post:
      tags: [Profile]
      summary: Change password
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/api-keys:
    get:
      tags: [API Keys]
      summary: List active API keys (no plaintext)
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Key metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKey"
    post:
      tags: [API Keys]
      summary: Create or rotate a named API key
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 80
      responses:
        "201":
          description: Key created; plaintext shown once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/APIError"
...
  /api/v1/tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
        - name: project
          in: query
          schema:
            type: string
          description: Project id, or none/0 for no project
        - name: tag
          in: query
          schema:
            type: string
        - name: due
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: completed
          in: query
          schema:
            type: string
        - name: workflow_claim_scope
          in: query
          schema:
            type: string
            enum: [mine, all]
          description: >
            When mine, hide unclaimed kanban-project tasks (classic tasks unchanged).
            Use all for the kanban board backlog. Default is no claim filter when omitted.
      responses:
        "200":
          description: Paginated task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
    post:
      tags: [Tasks]
      summary: Create a task
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/projects:
    get:
      tags: [Projects]
      summary: List projects
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Projects (owned first by position, then shared by name)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      tags: [Projects]
      summary: Create a project
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: List in-app notifications
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
...
  /api/v1/project-invites:
    get:
      tags: [Projects]
      summary: List pending invites for the current user
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Pending invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectInvite"
...
  /api/v1/share-links:
    get:
      tags: [Projects]
      summary: List share links for a scope
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - name: scope_type
          in: query
          required: true
          schema:
            type: string
            enum: [project, tag]
        - name: scope_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Share links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShareLink"
    post:
      tags: [Projects]
      summary: Create a read-only share link
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareLinkCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLink"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/tags:
    get:
      tags: [Tags]
      summary: List tags
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      tags: [Tags]
      summary: Create a tag (or return existing by name)
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameBody"
      responses:
        "201":
          description: Created or existing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/saved-views:
    get:
      tags: [Saved Views]
      summary: List saved views
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Saved views
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedView"
    post:
      tags: [Saved Views]
      summary: Create a saved view
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedViewCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedView"
        "409":
          $ref: "#/components/responses/APIError"
...
  /api/v1/dashboard:
    get:
      tags: [Dashboard]
      summary: Aggregate task stats for the current user
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
...
  /api/v1/calendar:
    get:
      tags: [Calendar]
      summary: Get or create the ICS feed token and URL
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Calendar token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarInfo"
...
  /api/v1/export:
    get:
      tags: [Export]
      summary: Export tasks as JSON or CSV
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        "200":
          description: Export payload
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
            text/csv:
              schema:
                type: string
...
  /api/v1/announcements/dismiss:
    post:
      tags: [Announcements]
      summary: Dismiss the global announcement banner for this session
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
  /api/v1/invites:
    get:
      tags: [Invites]
      summary: List invites
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invite"
    post:
      tags: [Invites]
      summary: Create an invite
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
...
  /api/v1/admin/settings:
    get:
      tags: [Admin]
      summary: Get site settings
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
    patch:
      tags: [Admin]
      summary: Update site settings
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsPatch"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
  /api/v1/admin/users:
    get:
      tags: [Admin]
      summary: List users
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: API key from profile, bootstrap, or device auth
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: httpOnly session cookie from /api/v1/auth/login or register
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    APIError:
      description: API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    OK:
      type: object
      properties:
        ok:
          type: boolean
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        user_name:
          type: string
        timezone:
          type: string
        items_per_page:
          type: integer
        permissions:
          type: array
          items:
            type: string
        digest_enabled:
          type: boolean
        digest_hour:
          type: integer
          minimum: 0
          maximum: 23
        allow_project_invites:
          type: boolean
          description: When false, the user cannot be invited to shared projects
        username_change_available:
          type: boolean
          description: When true, the user may call POST /api/v1/me/username once
    UserPatch:
      type: object
      properties:
        timezone:
          type: string
        items_per_page:
          type: integer
          enum: [10, 15, 25, 50]
        digest_enabled:
          type: boolean
        digest_hour:
          type: integer
          minimum: 0
          maximum: 23
        allow_project_invites:
          type: boolean
    RegisterRequest:
      type: object
      required: [email, password, confirm_password, user_name]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        confirm_password:
          type: string
        user_name:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[A-Za-z0-9_]+$'
          description: Unique username; immutable after registration except by admin
        timezone:
          type: string
        invite_token:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...
    DeviceCodeResponse:
      type: object
      properties:
        device_code:
          type: string
        user_code:
          type: string
        verification_uri:
          type: string
        verification_uri_complete:
          type: string
        expires_in:
          type: integer
        interval:
          type: integer
        redirect_uri:
          type: string
          description: Echo of the validated native deep link, when requested
    DeviceStatus:
      type: object
      properties:
        user_code:
          type: string
        client_name:
          type: string
        status:
          type: string
        redirect_uri:
          type: string
          description: Native deep link to open after a decision (when requested at code issuance)
    DeviceDecisionResponse:
      type: object
      properties:
        ok:
          type: boolean
        status:
          type: string
          description: approved or denied
        redirect_uri:
          type: string
          description: |
            Deep link for the browser to open. On approve includes status=approved;
            on deny includes error=access_denied. Omitted when no redirect_uri was requested.
...